Refactor with tidy workflow and jobs
