--- .github/chainguard/self.bazel-native-tidy.push-branch.sts.yaml
@@ -1,12 +1,0 @@
----
-issuer: https://token.actions.githubusercontent.com
-
-subject: repo:DataDog/datadog-agent:pull_request
-
-claim_pattern:
-  event_name: pull_request
-  ref: refs/pull/[0-9]+/merge
-  job_workflow_ref: DataDog/datadog-agent/\.github/workflows/bazel-native-tidy\.yml@refs/pull/[0-9]+/merge
-
-permissions:
-  contents: write
