--- .github/workflows/bazel-native-tidy.yml
@@ -1,54 +1,0 @@
----
-name: "Refresh Bazel http_archive sha256 on Renovate PRs"
-on:
-  pull_request:
-    types:
-      - labeled
-      - synchronize
-    paths:
-      - "deps/repos.MODULE.bazel"
-
-permissions: {}
-
-jobs:
-  refresh_archive_hashes:
-    if: ${{ github.repository == 'DataDog/datadog-agent' && github.event.pull_request.user.login == 'renovate[bot]' && contains(github.event.pull_request.labels.*.name, 'dependencies-bazel-native') }}
-    permissions:
-      id-token: write # Required for dd-octo-sts OIDC token
-    runs-on: ubuntu-latest
-    steps:
-      - uses: DataDog/dd-octo-sts-action@96a25462dbcb10ebf0bfd6e2ccc917d2ab235b9a # v1.0.4
-        id: octo-sts
-        with:
-          scope: DataDog/datadog-agent
-          policy: self.bazel-native-tidy.push-branch
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-        with:
-          ref: ${{ github.head_ref }}
-          fetch-depth: 0
-      - name: Install dda
-        uses: ./.github/actions/install-dda
-        with:
-          features: legacy-tasks
-      - name: Refresh http_archive sha256 for changed deps
-        run: dda inv -- bazel.refresh-archive-hashes --base-ref=origin/${{ github.event.pull_request.base.ref }}
-      - name: Create commit
-        id: commit
-        run: |
-          if git diff --quiet; then
-            echo "No changes to commit"
-            echo "has_changes=false" >> $GITHUB_OUTPUT
-          else
-            git config --global user.name "Login will be determined by the Github API based on the creator of the token"
-            git config --global user.email ""
-            git commit -am "[renovate skip] Auto-refresh http_archive sha256"
-            echo "has_changes=true" >> $GITHUB_OUTPUT
-          fi
-      - name: "Push signed commits"
-        if: steps.commit.outputs.has_changes == 'true'
-        uses: chouetz/push-signed-commits@9b123e8898e6176af65b9427b434e07aebe7f72f
-        with:
-          github-token: ${{ steps.octo-sts.outputs.token }}
-          local_branch_name: ${{ github.head_ref }}
-          remote_name: "origin"
-          remote_branch_name: ${{ github.head_ref }}
