+++ tasks/__init__.py
@@ -12,7 +12,6 @@ from tasks import (
     agent_ci_api,
     ami,
     auth,
-    bazel,
     bench,
     buildimages,
     claude,
@@ -200,7 +199,6 @@ ns.add_collection(auth)
 ns.add_collection(agent)
 ns.add_collection(ami)
 ns.add_collection(agent_ci_api)
-ns.add_collection(bazel)
 ns.add_collection(buildimages)
 ns.add_collection(claude)
 ns.add_collection(cluster_agent)
